requirement.txt
fastapi
uvicorn
keras
tensorflow
numpy
pillow
jinja2